requirements.txt
Django==3.2.9
django_rest_swagger==2.2.0
django-cors-headers
djangorestframework==3.13.1
drf_yasg==1.20.0
psycopg2==2.9.3
psycopg2_binary==2.9.2
django-request-logging==0.7.5
world==5.0.1
pyyaml